requirements.txt
streamlit
opencv-python-headless
numpy
pillow
requests
streamlit-geolocation